chore: rename ReacordButton to RecordAudioButton
